show category my class

+++ app/src/main/java/com/cious/learnhub/data/network/api/model/enrollments/Data.kt
@@ -23,7 +23,6 @@ data class Data(
     @SerializedName("categoryId")
     val categoryId: String,
 
-
     @SerializedName("level")
     val level: String,
 
@@ -52,7 +51,7 @@ data class Data(
     val totalDuration: String,
 
     @SerializedName("progress")
-    val progress: Int,
+    val progress: Double,
 
     @SerializedName("updatedAt")
     val updatedAt: String,
